move info.py
